requirements.txt
langchain-community
langchain-openai
faiss-cpu
langchain
python-dotenv